==> rutinas/02_clasificacion_variables/55_chaid_informe_wili.Rmd <==
---
title: CHAID - `r nombre_domest`
author: "DINEM"
output: 
  html_document:
    number_sections: true
params: 
  subpoblacion: subpoblacion
---


```{r setup , include=FALSE}
knitr::opts_chunk$set(
	echo = FALSE,
	message = FALSE,
	warning = FALSE,
	fig.height=10, fig.width=20, dpi=90, out.width=800
)
```

```{r library}
library(dplyr)
library(knitr) 
# library(CHAID)
library(partykit)
```


```{r datas}


# load("RData/enemdu_censo_desempleo.RData")
# Filtro a nivel de viviendas
aux <- pobreza_01 %>% 
  distinct(id_vivienda, .keep_all = T) %>% 
  mutate(depen1= case_when(depen==0 ~ 0, 
              depen>0 & depen<=1.25 ~ 1.25,
              depen>1 & depen<=2 ~ 2,
              depen>2 ~ 3))

aux_jef_hog<- pobreza_01 %>% 
  filter(parentesco==1)
aux_edc_bsc<- pobreza_01 %>% 
  filter(edad >= 5 & edad<= 14 )
aux_bch_ajus<- pobreza_01 %>% 
  filter(edad >= 15 & edad <= 17)
aux_alfabetismo<- pobreza_01 %>% 
  filter(edad>=15)
aux_t_infa<- pobreza_01 %>% 
  filter(edad>=5 & edad<=17)
aux_jov_no_est_tra<-pobreza_01 %>% 
  filter(edad >= 18 & edad <=24)
aux_ninios_muj<- pobreza_01 %>% 
  filter((edad<=5) | (sexo==2 & edad<=49 & edad>=15))


# Filtro a nivel de viviendas
# aux1 <- desempleo %>% 
#   distinct(id_vivienda, .keep_all = T) %>% 
#   mutate(depen1= case_when(depen==0 ~ 0, 
#               depen>0 & depen<=1.25 ~ 1.25,
#               depen>1 & depen<=2 ~ 2,
#               depen>2 ~ 3))
# 
# aux1_jef_hog <- desempleo %>% 
#   filter(parentesco==1)
# # aux1_edc_bsc<- desempleo %>% 
# #   filter(edad >= 5 & edad<= 14 )
# aux1_bch_ajus <- desempleo %>% 
#   filter(edad >= 15 & edad <= 17)
# aux1_alfabetismo <- desempleo %>% 
#   filter(edad>=15)
# aux1_t_infa <- desempleo %>% 
#   filter(edad>=5 & edad<=17)
# aux1_jov_no_est_tra <-desempleo %>% 
#   filter(edad >= 18 & edad <=24)
# aux1_niños_muj<- desempleo %>% 
#   filter((edad<=5) | (sexo==2 & edad<=49 & edad>=15))
```

# Pobreza por ingresos (1:Pobre, 0: No pobre)

## Techo de las viviendas

1: Hormigón (losa, cemento),  

2: Fibrocemento,asbesto (eternit, eurolit), 

3: Zinc, Aluminio, 

4: Teja, 

5: Palma, paja u hoja, 

6: Otro Material

```{r}
#Techo
# chaid_urb_rur <- ctree(pobreza_01 ~ techo+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ techo,
                   data = aux,
                   weights=fexp_cal_hog)


plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```


<!-- ## Techo y estado del techo -->

<!-- 1: Hormigón (losa, cemento),   -->

<!-- 2: Fibrocemento,asbesto (eternit, eurolit),  -->

<!-- 3: Zinc, Aluminio,  -->

<!-- 4: Teja,  -->

<!-- 5: Palma, paja u hoja,  -->

<!-- 6: Otro Material -->

<!-- 1: Bueno,  -->

<!-- 2: Bueno,  -->

<!-- 3: Malo. -->

<!-- ```{r} -->
<!-- #Techo -->
<!-- chaid <- ctree(pobreza_01 ~ techo+estado_techo, -->
<!--                    data = aux, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->



## Piso

1: Duela, parquet, tablón tratado o piso flotante, 

2: Cerámica, baldosa, vinil o porcelanato, 

3: Mármol o marmetón, 

4: Ladrillo o cemento, 

5: Tabla / tablón no tratado, 

6: Caña, 

7: Tierra, 

8: Otro Material.

```{r}
# #Piso
# chaid_urb_rur <- ctree(pobreza_01 ~ piso+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ piso,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```


<!-- ## Piso y estado del piso -->

<!-- 1: Duela, parquet, tablón tratado o piso flotante,  -->

<!-- 2: Cerámica, baldosa, vinil o porcelanato,  -->

<!-- 3: Mármol o marmetón,  -->

<!-- 4: Ladrillo o cemento,  -->

<!-- 5: Tabla / tablón no tratado,  -->

<!-- 6: Caña,  -->

<!-- 7: Tierra,  -->

<!-- 8: Otro Material. -->

<!-- 1: Bueno,  -->

<!-- 2: Bueno,  -->

<!-- 3: Malo. -->

<!-- ```{r} -->
<!-- #Piso -->

<!-- chaid <- ctree(pobreza_01 ~ estado_piso+piso, -->
<!--                    data = aux, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->


## Paredes

1: Hormigón/Ladrillo o Bloque, 

2: Asbesto/Cemento (Fibrolit), 

3: Adobe o Tapia, 

4: Madera, 

5: Caña revestida o bahareque, 

6: Caña no revestida o estera, 

7: Otra Material

```{r }
#Paredes exteriores
# chaid_urb_rur <- ctree(pobreza_01 ~ paredes+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ paredes,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))

```


<!-- ## Paredes y estado de las paredes -->

<!-- 1: Hormigón/Ladrillo o Bloque,  -->

<!-- 2: Asbesto/Cemento (Fibrolit),  -->

<!-- 3: Adobe o Tapia,  -->

<!-- 4: Madera,  -->

<!-- 5: Caña revestida o bahareque,  -->

<!-- 6: Caña no revestida o estera,  -->

<!-- 7: Otra Material -->

<!-- 1: Bueno,  -->

<!-- 2: Bueno,  -->

<!-- 3: Malo. -->

<!-- ```{r} -->
<!-- #Piso -->
<!-- chaid <- ctree(pobreza_01 ~ estado_paredes+paredes, -->
<!--                    data = aux, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->

## Procedencia del agua

1: Red pública, 

2: Pila o llave pública, 

3: Otra fuente por tubería, 

4: Carro repartidor, triciclo, 

5: Pozo, 

6: Río, vertiente, acequia, 

7: Otro

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ p_agua+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ p_agua,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```


## Tubería por dentro del domicilio

1: Por tubería dentro de la vivienda, 

2: Por tubería fuera de la vivienda pero en el lote, 

3: Por tubería fuera de la vivienda, lote o terreno, 

4: No recibe agua por tubería sino por otros medios

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ r_agua+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ r_agua,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```


## Servicio Higiénico

1: Excusado y alcantarillado, 

2: Excusado y pozo séptico, 

3: Excusado y pozo ciego, 

4: Letrina, 

5: No tiene.

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ tipo_sh+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ tipo_sh,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```


## Luz

1: Empresa eléctrica pública, 

2: Planta eléctrica privada, 

3: Vela, candil, mechero, gas, 

4: Ninguno.

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ luz+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ luz,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```


## Basura

1: Contratan el servicio, 

2: Servicio municipal, 

3: Botan a la calle, quebrada, río, 

4: La queman, entierran, 

5: Otra.

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ basura+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ basura,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```

## Hacinamiento

0: Hacinado   

1: No Hacinado

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ hacinam+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ hacinam,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```


## Dependencia


```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ depen1+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ depen1,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```


## Jefes de hogar Escolaridad

1: >13  
 
0: <=13
 
```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ esc_ea+area,
#                    data = aux_jef_hog ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ esc_ea,
                   data = aux_jef_hog,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```


## Asistencia de educación general básica (5 a 14 años)

1: Asiste

2: No asiste 

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ asiste_clases+area,
#                    data = aux_edc_bsc ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ asiste_clases,
                   data = aux_edc_bsc,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```


## Asistencia ajustada en bachillerato(15 a 17 años))
 
1: Asiste
 
2: No asiste 


```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ asiste_clases+area,
#                    data = aux_bch_ajus ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ asiste_clases,
                   data = aux_bch_ajus,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```


## Alfabetismo
1: sabe leer y escribir
 
0: no sabe leer y escribir

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ leer_escribir+area,
#                    data = aux_alfabetismo ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ leer_escribir,
                   data = aux_alfabetismo,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```

## Combustible
 1: Gas,  
 
 2: Leña/carbón,   
 
 3: Electricidad , 
 
 4: Otro

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ combustible+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))

chaid <- ctree(pobreza_01 ~ combustible,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```

## Espacio exclusivo para cocinar

 1: Si,  
 
 2: No

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ exc_cocina+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))

chaid <- ctree(pobreza_01 ~ exc_cocina,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```


## Ducha

 1: Exclusivo del hogar,  
 
 2: Compartido con otros hogares,  
 
 3: No tiene

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ exc_ducha+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
#   plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))

chaid <- ctree(pobreza_01 ~ exc_ducha,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```

## Seguro

```{r}
#Techo
# chaid_urb_rur <- ctree(pobreza_01 ~ seguros_1+seguros_2+area,
#                    data = pobreza_01 ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ seguros_1+seguros_2,
                   data = pobreza_01,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```


## Autoidentificación

1: Indígena, 

2: Afroecuatoriano, 

3: Negro, 

4: Mulato, 

5: Montubio, 

6: Mestizo, 

7: Blanco, 

8: Otro, cual.

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ autoidentifica+area,
#                    data = pobreza_01 ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ autoidentifica,
                   data = pobreza_01,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```



## Seguros

1: SG, 

2: Voluntario, 

3: Campesiono , 

4: ISSFA/ISSPOL, 

5: SPhosp, 

6: SP_sn_hosp, 

7: AUS, 

8: Municipales_CP, 

9: MSP, 

10: Ninguno.

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ seguros_1 +seguros_2+area,
#                    data = pobreza_01 ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ seguros_1 +seguros_2,
                   data = pobreza_01,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```



## Trabajo infantil

1: Trabajo  

2: No Trabajo

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~trabajo+area,
#                    data = aux_t_infa ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ trabajo,
                   data = aux_t_infa,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```

## Jóvenes que ni trabajan ni estudian

0: Trabajo  o estudia

1: Ni Trabaja ni estudia

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~nini+area,
#                    data = aux_jov_no_est_tra ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ nini,
                   data = aux_jov_no_est_tra,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```

## Hogares monoparentales cuyo jefe de hogar es mujer y tiene hijos.

1: Hogar monoparental cuyo jefe de hogar es mujer y tiene hijos

0: Otros

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~id_hmuj+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ id_hmuj,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```

## Relación niños mujeres 

```{r}

# chaid_urb_rur <- ctree(pobreza_01 ~ni_mj+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ ni_mj,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```

## Porcentaje de mujeres y Porcentaje de hombres

1: Hombres

2: Mujeres

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~sexo+area,
#                    data = aux ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ sexo,
                   data = aux,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```

## Tasa de juventud

1: Personas entre 15 y 29 años

0: Otros

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ta_juventud+area,
#                    data = pobreza_01 ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ ta_juventud,
                   data = pobreza_01,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```

## Tasa de Envejecimiento

1: Personas mayores a 65 años 

0: Otros

```{r}
# chaid_urb_rur <- ctree(pobreza_01 ~ta_envejecimiento+area,
#                    data = pobreza_01 ,
#                    weights=fexp_cal_hog)
# 
# plot(chaid_urb_rur,
#      main = "Urbano - Rural",
#      gp = gpar(fontsize = 13))


chaid <- ctree(pobreza_01 ~ ta_envejecimiento,
                   data = pobreza_01,
                   weights=fexp_cal_hog)

plot(chaid,
     main = "Dominio estratificación",
     gp = gpar(fontsize = 13))
```

<!-- # Empleo inadecuado y desempleo (1: , 0: empleo adecuado) -->

<!-- ## Techo de las viviendas -->

<!--  1: Hormigón (losa, cemento),  2: Fibrocemento,asbesto (eternit, eurolit), 3: Zinc, Aluminio, 4: Teja, 5: Palma, paja u hoja, 6: Otro Material -->


<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ techo+area, -->
<!--                        data = aux1 , -->
<!--                        weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ techo, -->
<!--                data = aux1, -->
<!--                weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->

<!-- ## Techo y estado del techo -->

<!-- * 1: Hormigón (losa, cemento),  2: Fibrocemento,asbesto (eternit, eurolit), 3: Zinc, Aluminio, 4: Teja, 5: Palma, paja u hoja, 6: Otro Material -->

<!-- * 1: Bueno, 2: Bueno, 3: Malo. -->

<!-- ```{r} -->
<!-- #Techo -->
<!-- chaid <- ctree(prec_laboral ~ techo+estado_techo, -->
<!--                    data = aux1, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->

<!-- ## Piso -->

<!-- 1: Duela, parquet, tablón tratado o piso flotante, 2: Cerámica, baldosa, vinil o porcelanato, 3: Mármol o marmetón, 4: Ladrillo o cemento, 5: Tabla / tablón no tratado, 6: Caña, 7: Tierra, 8: Otro Material. -->

<!-- ```{r} -->
<!-- #Piso -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ piso+area, -->
<!--                        data = aux1 , -->
<!--                        weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ piso, -->
<!--                data = aux1, -->
<!--                weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->


<!-- ## Piso y estado del piso -->

<!-- * 1: Duela, parquet, tablón tratado o piso flotante, 2: Cerámica, baldosa, vinil o porcelanato, 3: Mármol o marmetón, 4: Ladrillo o cemento, 5: Tabla / tablón no tratado, 6: Caña, 7: Tierra, 8: Otro Material. -->

<!-- * 1: Bueno, 2: Bueno, 3: Malo. -->

<!-- ```{r} -->
<!-- #Piso -->

<!-- chaid <- ctree(prec_laboral ~ estado_piso+piso, -->
<!--                    data = aux1, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->


<!-- ## Paredes -->

<!-- 1: Hormigón/Ladrillo o Bloque, 2: Asbesto/Cemento (Fibrolit), 3: Adobe o Tapia, 4: Madera, 5: Caña revestida o bahareque, 6: Caña no revestida o estera, 7: Otra Material -->

<!-- ```{r } -->
<!-- #Paredes exteriores -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ paredes+area, -->
<!--                        data = aux1 , -->
<!--                        weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ paredes, -->
<!--                data = aux1, -->
<!--                weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->

<!-- ``` -->

<!-- ## Paredes y estado de las paredes -->

<!-- * 1: Hormigón/Ladrillo o Bloque, 2: Asbesto/Cemento (Fibrolit), 3: Adobe o Tapia, 4: Madera, 5: Caña revestida o bahareque, 6: Caña no revestida o estera, 7: Otra Material -->

<!-- * 1: Bueno, 2: Bueno, 3: Malo. -->

<!-- ```{r} -->
<!-- #Piso -->
<!-- chaid <- ctree(prec_laboral ~ estado_paredes+paredes, -->
<!--                    data = aux1, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->

<!-- ## Procedencia del agua -->

<!-- 1: Red pública, 2: Pila o llave pública, 3: Otra fuente por tubería, 4: Carro repartidor, triciclo, 5: Pozo, 6: Río, vertiente, acequia, 7: Otro -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ prov_agua+area, -->
<!--                        data = aux1 , -->
<!--                        weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ prov_agua, -->
<!--                data = aux1, -->
<!--                weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->


<!-- ## Tubería por dentro del domicilio -->

<!-- 1: Por tubería dentro de la vivienda, 2: Por tubería fuera de la vivienda pero en el lote, 3: Por tubería fuera de la vivienda, lote o terreno, 4: No recibe agua por tubería sino por otros medios -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ agua+area, -->
<!--                        data = aux1 , -->
<!--                        weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ agua, -->
<!--                data = aux1, -->
<!--                weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->


<!-- ## Servicio Higiénico -->

<!-- 1: Excusado y alcantarillado, 2: Excusado y pozo séptico, 3: Excusado y pozo ciego, 4: Letrina, 5: No tiene. -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ sh+area, -->
<!--                        data = aux1 , -->
<!--                        weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ sh, -->
<!--                data = aux1, -->
<!--                weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->

<!-- ## Luz -->

<!-- 1: Empresa eléctrica pública, 2: Planta eléctrica privada, 3: Vela, candil, mechero, gas, 4: Ninguno. -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ luz+area, -->
<!--                    data = aux1 , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ luz, -->
<!--                    data = aux1, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->




<!-- ## Basura -->

<!-- 1: Contratan el servicio, 2: Servicio municipal, 3: Botan a la calle, quebrada, río, 4: La queman, entierran, 5: Otra. -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ basura+area, -->
<!--                        data = aux1 , -->
<!--                        weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ basura, -->
<!--                data = aux1, -->
<!--                weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->

<!-- ## Hacinamiento -->

<!-- 0: Hacinado   1: No Hacinado -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ hacinam+area, -->
<!--                    data = aux1 , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ hacinam, -->
<!--                    data = aux1, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->

<!-- ## Dependencia -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ depen1+area, -->
<!--                    data = aux1 , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ depen1, -->
<!--                    data = aux1, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->



<!-- ## Jefes de hogar Escolaridad -->

<!--  1: >13   -->
<!--  0: <=13 -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ esc_ea+area, -->
<!--                    data = aux1_jef_hog , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ esc_ea, -->
<!--                    data = aux1_jef_hog, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->

<!-- ## Asistencia de ajustada en bachillerato(15 a 17 años)) -->

<!--  1: Asiste -->
<!--  0: No asiste  -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ asiste_clases+area, -->
<!--                    data = aux1_bch_ajus , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ asiste_clases, -->
<!--                    data = aux1_bch_ajus, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->


<!-- ## Alfabetismo -->
<!--  1: Asiste -->
<!--  0: No asiste  -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ leer_escribir+area, -->
<!--                    data = aux1_alfabetismo , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ leer_escribir, -->
<!--                    data = aux1_alfabetismo, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->

<!-- ## Combustible -->
<!--  1: Gas,  2: Leña/carbón,   3: Electricidad , 4: Otro -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ mat_cocinan+area, -->
<!--                    data = aux1 , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->

<!-- chaid <- ctree(prec_laboral ~ mat_cocinan, -->
<!--                    data = aux1, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->


<!-- ## Espacio exclusivo para cocinar -->
<!--  1: Si,  2: No -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ exc_cocina+area, -->
<!--                    data = aux1 , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->

<!-- chaid <- ctree(prec_laboral ~ exc_cocina, -->
<!--                    data = aux1, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->


<!-- ## Ducha -->
<!--  1: Exclusivo del hogar,  2: Compartido con otros hogares,   3: No tiene -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ exc_ducha+area, -->
<!--                    data = aux1 , -->
<!--                    weights=fexp_cal_hog) -->

<!--   plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->

<!-- chaid <- ctree(prec_laboral ~ exc_ducha, -->
<!--                    data = aux1, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->


<!-- ## Autoidentificación -->

<!-- 1: Indígena, 2: Afroecuatoriano, 3: Negro, 4: Mulato, 5: Montubio, 6: Mestizo, 7: Blanco, 8: Otro, cual. -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ autoidentif+area, -->
<!--                        data = desempleo , -->
<!--                        weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ autoidentif, -->
<!--                data = desempleo, -->
<!--                weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->

<!-- ## Seguros -->

<!-- 1: SG, 2: Voluntario, 3: Campesiono , 4: ISSFA/ISSPOL, 5: SPhosp, 6: SP_sn_hosp, 7: AUS, 8: Municipales_CP, 9: MSP, 10: Ninguno. -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ ss_1+ss_2+area, -->
<!--                    data = desempleo , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ ss_1+ss_2, -->
<!--                    data = desempleo, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->


<!-- ## Trabajo infantil -->

<!-- 1: Trabajo   -->
<!-- 2: No Trabajo -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~trabajo+area, -->
<!--                    data = aux1_t_infa , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ trabajo, -->
<!--                    data = aux1_t_infa, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->

<!-- ## Jóvenes que ni trabajan ni estudian -->

<!-- 1: Trabajo   -->
<!-- 2: No Trabajo -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~nini+area, -->
<!--                    data = aux1_jov_no_est_tra , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ nini, -->
<!--                    data = aux1_jov_no_est_tra, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->

<!-- ## Hogares monoparentales cuyo jefe de hogar es mujer y tiene hijos. -->

<!-- 1: Hogar monoparental cuyo jefe de hogar es mujer y tiene hijos -->
<!-- 0: Otros -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~id_hmuj+area, -->
<!--                    data = aux1 , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ id_hmuj, -->
<!--                    data = aux1, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->


<!-- ## Relación niños mujeres  -->

<!-- ```{r} -->
<!-- # chaid_urb_rur <- ctree(prec_laboral ~ni_mj+area, -->
<!-- #                    data = aux1 , -->
<!-- #                    weights=fexp_cal_hog) -->
<!-- #  -->
<!-- # plot(chaid_urb_rur, -->
<!-- #      main = "Urbano - Rural", -->
<!-- #      gp = gpar(fontsize = 13)) -->

<!-- chaid <- ctree(prec_laboral ~ ni_mj, -->
<!--                    data = aux1, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->


<!-- ## Porcentaje de mujeres y dim8_var2: Porcentaje de hombres -->

<!-- 1: Hombres -->
<!-- 2: Mujeres -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~sexo+area, -->
<!--                    data = aux1 , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ sexo, -->
<!--                    data = aux1, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->

<!-- ## Tasa de juventud -->
<!-- 1: Personas entre 15 y 29 años -->
<!-- 0: Otros -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ta_juventud+area, -->
<!--                    data = desempleo , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ ta_juventud, -->
<!--                    data = desempleo, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->

<!-- ## Tasa de Envejecimiento -->
<!-- 1: Personas mayores a 65 años  -->
<!-- 0: Otros -->

<!-- ```{r} -->
<!-- chaid_urb_rur <- ctree(prec_laboral ~ta_envejecimiento+area, -->
<!--                    data = desempleo , -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid_urb_rur, -->
<!--      main = "Urbano - Rural", -->
<!--      gp = gpar(fontsize = 13)) -->


<!-- chaid <- ctree(prec_laboral ~ ta_envejecimiento, -->
<!--                    data = desempleo, -->
<!--                    weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->



<!-- # Combinación de varias variables de vivienda -->

<!-- ```{r} -->
<!-- chaid <- ctree(prec_laboral ~ techo+paredes+piso, -->
<!--                data = aux1 , -->
<!--                weights=fexp_cal_hog) -->

<!-- plot(chaid, -->
<!--      main = "Dominio estratificación", -->
<!--      gp = gpar(fontsize = 13)) -->
<!-- ``` -->
